requirements.txt
coverage~=7.6.9
Django~=4.2.17
django-stubs~=5.1.1
djangorestframework~=3.15.2
djangorestframework-stubs~=3.15.2
mypy~=1.13.0
mypy-extensions~=1.0.0
pre-commit
pytest~=8.3.4
pytest-django~=4.9.0
